requirements.txt
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.65.0
git+https://github.com/openai/CLIP.git
numpy==1.24.3
Pillow==9.5.0
decord==0.6.0
gradio==3.23.0
markdown2==2.4.8
einops==0.6.1
requests==2.30.0
sentencepiece==0.1.99
protobuf==4.23.2
accelerate==0.20.3
tokenizers>=0.13.3
pydantic==1.10.7
git+https://github.com/m-bain/whisperx.git
git+https://github.com/shehanmunasinghe/whisper-at.git@patch-1#subdirectory=package/whisper-at
git+https://github.com/xinyu1205/recognize-anything.git
transformers@git+https://github.com/huggingface/transformers.git@cae78c46
openai==0.28.0
scenedetect[opencv-headless]